+++ src/bedrock/world/simple_sparse_container.h
@@ -14,27 +14,9 @@
 
 #pragma once
 
-#include <cstdint>
+#include "bedrock/world/container.h"
 
-enum class ItemStackRequestActionType : std::uint8_t {
-    Take = 0,
-    Place = 1,
-    Swap = 2,
-    Drop = 3,
-    Destroy = 4,
-    Consume = 5,
-    Create = 6,
-    PlaceInItemContainerDeprecated = 7,
-    TakeFromItemContainerDeprecated = 8,
-    ScreenLabTableCombine = 9,
-    ScreenBeaconPayment = 10,
-    ScreenHUDMineBlock = 11,
-    CraftRecipe = 12,
-    CraftRecipeAuto = 13,
-    CraftCreative = 14,
-    CraftRecipeOptional = 15,
-    CraftRepairAndDisenchant = 16,
-    CraftLoom = 17,
-    CraftNonImplementedDeprecated = 18,
-    CraftResultsDeprecated = 19,
+class SimpleSparseContainer : public Container, public ContainerContentChangeListener {
+public:
+    ~SimpleSparseContainer() override = 0;
 };